pyproject.toml
[tool.poetry]
name = "auth-fastapi"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
python-dotenv = "^1.0.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {version = "1.7.4", extras = ["bcrypt"]}
python-multipart = "^0.0.9"
pyjwt = "^2.8.0"
sqlalchemy = "^2.0.31"


bcrypt = "4.1.2"
pydantic-settings = "^2.3.4"
black = "^24.4.2"
flake8 = "^7.1.0"
isort = "^5.13.2"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79

[tool.poetry.scripts]
seed_users = "scripts.seed_users:main"
